requirements.txt
SQLAlchemy~=1.4.41
PyJWT~=2.4.0
Flask~=2.2.2
requests~=2.28.1